Struct workers::Workers

pub struct Workers { /* private fields */ }

Manages a collection of workers.

Implementations

impl Workers

pub async fn from_path( workers_dir: PathBuf, key_storage: Arc<KeyStorage>, core_manager: Arc<CoreManager>, channel_size: usize ) -> Result<(Self, Receiver<Event>)>

Creates a Workers instance by loading persisted worker data from the specified directory.

Arguments

workers_dir - The path to the directory containing persisted worker data.

key_storage - An Arc<KeyStorage> instance for managing worker key pairs.

scope - A Scope instance used to determine the host and manage key pairs.

Returns

Returns Result<Self, eyre::Error> where:

Ok(workers) if the Workers instance is successfully created.

Err(eyre::Error) if an error occurs during the creation process.

pub fn get_deal_id(&self, worker_id: WorkerId) -> Result<DealId, WorkersError>

Retrieves the deal ID associated with the specified worker ID.

Arguments

worker_id - The PeerId of the worker for which the deal ID is requested.

Returns

Returns Result<DealId, WorkersError> where:

Ok(deal_id) if the deal ID is successfully retrieved.

Err(WorkersError) if an error occurs, such as the worker not found.

pub async fn create_worker( &self, params: WorkerParams ) -> Result<WorkerId, WorkersError>

Creates a new worker with the given deal_id and initial peer ID.

Arguments

deal_id - A String representing the unique identifier for the deal associated with the worker.

init_peer_id - The initial PeerId of the worker.

Returns

Returns Result<PeerId, WorkersError> where:

Ok(worker_id) if the worker is successfully created, returning the ID of the created worker.

Err(WorkersError) if an error occurs, such as the worker already existing or key pair creation failure.

pub async fn remove_worker( &self, worker_id: WorkerId ) -> Result<(), WorkersError>

Removes a worker with the specified worker_id.

Arguments

worker_id - The PeerId of the worker to be removed.

Returns

Returns Result<(), WorkersError> where:

Ok(()) if the worker is successfully removed.

Err(WorkersError) if an error occurs, such as the worker not found or key pair removal failure.

pub async fn activate_worker( &self, worker_id: WorkerId ) -> Result<(), WorkersError>

Activates the worker with the specified worker_id.

The activation process sets the worker’s status to true, indicating that the worker is active. The updated status is persisted, and internal data structures are updated.

Arguments

worker_id - The PeerId of the worker to be activated.

Returns

Returns Result<(), WorkersError> where:

Ok(()) if the activation is successful.

Err(WorkersError) if an error occurs during the activation process.

pub async fn deactivate_worker( &self, worker_id: WorkerId ) -> Result<(), WorkersError>

Deactivates the worker with the specified worker_id.

The deactivation process sets the worker’s status to false, indicating that the worker is not active. The updated status is persisted, and internal data structures are updated.

Arguments

worker_id - The PeerId of the worker to be deactivated.

Returns

Returns Result<(), WorkersError> where:

Ok(()) if the deactivation is successful.

Err(WorkersError) if an error occurs during the deactivation process.

pub fn get_runtime_handle(&self, worker_id: WorkerId) -> Option<Handle>

pub fn get_worker_creator( &self, worker_id: WorkerId ) -> Result<PeerId, WorkersError>

Retrieves the creator PeerId associated with the specified worker PeerId.

If the provided worker_id belongs to the host, the host’s PeerId is returned. Otherwise, the creator’s PeerId associated with the worker is retrieved.

Arguments

worker_id - The PeerId of the worker for which the creator PeerId is requested.

Returns

Returns Result<PeerId, WorkersError> where:

Ok(creator_peer_id) if the creator PeerId is successfully retrieved.

Err(WorkersError) if an error occurs, such as the worker not found.

pub fn get_worker_id(&self, deal_id: DealId) -> Result<WorkerId, WorkersError>

Retrieves the worker ID associated with the specified deal_id.

Arguments

deal_id - The unique identifier (String) of the deal associated with the worker.

Returns

Returns Result<PeerId, WorkersError> where:

Ok(worker_id) if the worker ID is successfully retrieved.

Err(WorkersError) if an error occurs, such as the worker not found.

pub fn is_worker_active(&self, worker_id: WorkerId) -> bool

Checks the activation status of the worker with the specified worker_id.

The activation status indicates whether the worker is currently active or not. If the worker is the host, it is always considered active. Otherwise, the method checks the internal data structures to determine the worker’s status.

Arguments

worker_id - The PeerId of the worker to check for activation status.

Returns

Returns true if the worker is active, and false otherwise. If the worker with the specified worker_id is not found, a warning is logged, and false is returned.

pub fn list_workers(&self) -> Vec<WorkerId>

Retrieves a list of all worker IDs.

Returns

Returns a Vec<WorkerId> representing a list of all worker IDs currently registered.

pub fn shutdown(&self)